split cli (generator added)
